testing mounting of popups
