+++ package.json
@@ -15,7 +15,7 @@
 	"devDependencies": {
 		"@codemirror/lang-markdown": "^6.1.1",
 		"@codemirror/language-data": "^6.3.1",
-		"@floating-ui/dom": "^1.2.9",
+		"@floating-ui/dom": "^1.4.1",
 		"@skeletonlabs/skeleton": "^1.6.2",
 		"@sveltejs/adapter-auto": "^2.0.0",
 		"@sveltejs/kit": "^1.5.0",
