+++ src/lib/pages/admin/Editor/editor.d.ts
@@ -1,0 +1,9 @@
+interface ToastSettingsMap {
+	[key: string]: ToastSettings
+}
+
+interface PopupSettingsMap {
+	[key: string]: PopupSettings
+}
+
+export { ToastSettingsMap, PopupSettingsMap }
