+++ src/routes/+layout.svelte
@@ -14,8 +14,9 @@
 	import Footer from '../components/Layout/Footer/Footer.svelte'
 	import hljs from 'highlight.js'
 	import 'highlight.js/styles/github-dark.css'
+	import { storeHighlightJs } from '@skeletonlabs/skeleton'
 	import { computePosition, autoUpdate, offset, shift, flip, arrow } from '@floating-ui/dom'
-	import { storePopup, storeHighlightJs } from '@skeletonlabs/skeleton'
+	import { storePopup } from '@skeletonlabs/skeleton'
 	storePopup.set({ computePosition, autoUpdate, offset, shift, flip, arrow })
 	storeHighlightJs.set(hljs)
